requirements.txt
# Cat Classification Project - Requirements
# Python >= 3.8

# Deep Learning Framework
tensorflow>=2.13.0,<3.0.0
keras>=2.13.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
ultralytics>=8.0.0  # YOLO models

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Profiling & EDA
ydata-profiling>=4.5.0

# Utilities
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# For Google Colab (optional)
google-colab  # Only needed if running on Colab

# Export/Import
h5py>=3.9.0  # For Keras model saving

# Optional: GPU Support
# tensorflow-gpu>=2.13.0  # Use if you have NVIDIA GPU with CUDA

# Development Tools (optional)
# black>=23.0.0  # Code formatting
# flake8>=6.0.0  # Linting
# pytest>=7.4.0  # Testing
